Reorg UniformsUI folder
