+++ MetalGemini/UniformUI/VerticalSlider.swift
@@ -16,17 +16,18 @@ struct VerticalSlider: View {
         GeometryReader { geometry in
             ZStack(alignment: .bottom) {
                 Rectangle() // Slider Track
-                    .fill(Color.white.opacity(0.15))
+                    .fill(Color.white.opacity(0.25))
                     .frame(width: 40)
 
                 Rectangle() // Slider Thumb
-                    .fill(Color.init(red: 0, green: 0.6, blue: 0.85).opacity(0.7))
+                    .fill(Color.init(red: 0, green: 0.65, blue: 0.85).opacity(0.7))
                     .frame(width: 40, height: max(0,CGFloat((value - range.lowerBound) / (range.upperBound - range.lowerBound)) * geometry.size.height))
 
                 Rectangle() // Invisible Interactive Layer
                     .fill(Color.clear)
                     .contentShape(Rectangle()) // Ensure the entire area is interactive
             }
+            .cornerRadius(6)
             .gesture(DragGesture(minimumDistance: 0).onChanged { gesture in
                 let y = gesture.location.y
                 // Convert y position to a value in the range
